multiple fixes & polish
Fixed a typo in the unbound "Studies on the Wither" codex name, previously had a . at the end of the name.
Added a codex sprite for future unbound codices
added a dark border to the icons of the krythers study codices for consistency
fixed spellslinger after neb broke it (love you no homo)
fixed the rifter helmet small fossil typo in the treasurepool patch, thanks to Mediant.