Biome changes & misc
Made it so the tidal remnants is considered a mini-biome and not a secondary region anymore
changed saltshaker object to drop recipe instead of gun (for balance)
added unbound village guard with dialogue
added wip structures for the unbound village
it probably says i updated the mission 3. it is a lie, i tried to work on it but as soon as i removed a few tiles i got discouraged and closed tiled. i pushed the changes tho