Update Avian Journal
Was a bit unsure if the last page meant they hastily scribbled something ineligible or if they said 'screw it' and scribbled over it, but I decided on the latter. I didn't proofread this in-game, but there should be enough space for things not to clip out of the box. I'm pretty sure ERM doesn't change the height of the pages